10-7959a | Update schema definitions (#39782)
