+++ src/applications/ivc-champva/10-7959a/tests/e2e/fixtures/data/no-packet.json
@@ -31,20 +31,20 @@
     "applicantMemberNumber": "123123123",
     "applicantName": {
       "first": "Beneficiary",
-      "middle": "Terry",
+      "middle": "T",
       "last": "Jones",
       "suffix": "II"
     },
     "applicantDOB": "2022-03-23",
     "sponsorName": {
       "first": "Sponsor",
-      "middle": "Quincy",
+      "middle": "Q",
       "last": "Jones",
       "suffix": "Sr."
     },
     "certifierRole": "applicant",
     "certifierReceivedPacket": false,
-    "statementOfTruthSignature": "Beneficiary Terry Jones",
+    "statementOfTruthSignature": "Beneficiary T Jones",
     "statementOfTruthCertified": true
   }
 }
