+++ src/applications/ivc-champva/10-7959a/tests/e2e/fixtures/data/test-data.json
@@ -16,7 +16,7 @@
     },
     "certifierName": {
       "first": "Jerry",
-      "middle": "Middle",
+      "middle": "M",
       "last": "Jones",
       "suffix": "Jr."
     },
@@ -25,7 +25,7 @@
     "claimStatus": "new",
     "sponsorName": {
       "first": "Sponsor",
-      "middle": "Quincy",
+      "middle": "Q",
       "last": "Jones",
       "suffix": "Jr."
     },
