+++ src/applications/ivc-champva/10-7959a/tests/e2e/fixtures/data/third-party-foreign-address-ohi-medical-claim-work-auto.json
@@ -37,14 +37,14 @@
     "applicantMemberNumber": "123123123",
     "applicantName": {
       "first": "Jim",
-      "middle": "Middle",
+      "middle": "M",
       "last": "Lastname",
       "suffix": "II"
     },
     "applicantDOB": "2022-05-14",
     "sponsorName": {
       "first": "John",
-      "middle": "Middle",
+      "middle": "M",
       "last": "Jones",
       "suffix": "Sr."
     },
@@ -62,7 +62,7 @@
     },
     "certifierName": {
       "first": "Jerry",
-      "middle": "Middle",
+      "middle": "M",
       "last": "Jones",
       "suffix": "Jr."
     },
@@ -82,7 +82,7 @@
         "providerPhone": "1231231235"
       }
     ],
-    "statementOfTruthSignature": "Certifier Middle Jones",
+    "statementOfTruthSignature": "Certifier M Jones",
     "statementOfTruthCertified": true
   }
 }
