+++ src/applications/ivc-champva/10-7959a/tests/e2e/fixtures/data/two-ohi-other-type.json
@@ -30,14 +30,14 @@
     "applicantMemberNumber": "321321321",
     "applicantName": {
       "first": "Beneficiary",
-      "middle": "Middle",
+      "middle": "M",
       "last": "Jones",
       "suffix": "II"
     },
     "applicantDOB": "2018-01-22",
     "sponsorName": {
       "first": "Sponsor",
-      "middle": "Middle",
+      "middle": "M",
       "last": "Jones",
       "suffix": "II"
     },
@@ -59,6 +59,6 @@
       }
     ],
     "statementOfTruthCertified": true,
-    "statementOfTruthSignature": "Beneficiary Middle Jones"
+    "statementOfTruthSignature": "Beneficiary M Jones"
   }
 }
